requirements.txt
ccxt
pandas
pandas-ta
python-telegram-bot==20.8
ollama